+++ headers/minirt.h
@@ -6,7 +6,7 @@
 /*   By: daprovin <daprovin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/22 17:33:39 by daprovin          #+#    #+#             */
-/*   Updated: 2020/02/07 18:20:02 by daprovin         ###   ########.fr       */
+/*   Updated: 2020/02/10 14:22:53 by daprovin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,8 +35,8 @@ typedef struct	s_vct
 
 typedef struct	s_res
 {
-	int				x;
-	int				y;
+	double			x;
+	double			y;
 }				t_res;
 
 typedef struct	s_ray
